+++ lib/genesmanproxy.dart
@@ -1,2 +1,4 @@
+export 'src/data_source.dart';
+export 'src/down_stream.dart';
 export 'src/download_meta.dart';
 export 'src/streamproxy.dart';